[KH] updating vim command to not always run logging  start time

+++ fish/config.fish
@@ -26,11 +26,11 @@ set -gx TERM xterm-256color
 source ~/.config/Juliet/fish/_util.fish
 
 function envsource
-  for line in (cat $argv | grep -v '^#' | grep -v '^\s*$')
-    set item (string split -m 1 '=' $line)
-    set -gx $item[1] $item[2]
-    echo "Exported key $item[1]"
-  end
+    for line in (cat $argv | grep -v '^#' | grep -v '^\s*$')
+        set item (string split -m 1 '=' $line)
+        set -gx $item[1] $item[2]
+        echo "Exported key $item[1]"
+    end
 end
 
 function p
@@ -132,8 +132,14 @@ function watch-repo
 end
 
 function vim
-    rm ~/vim.log
-    nvim --startuptime ~/vim.log $argv
+    argparse --name="debug" d/debug -- $argv
+
+    if test -n "$_flag_d"
+        rm ~/vim.log
+        nvim --startuptime ~/vim.log $argv
+    else
+        nvim $argv
+    end
 end
 
 function fish_greeting
